Création du hook pour récup les donnée depuis un form react pour l'envoyer au backend traiter par flask puis affichage sur une autre page react.
