+++ src/pages/Recap.jsx
@@ -10,6 +10,7 @@ export default function Recap() {
         <h2 className="text-xl font-bold mb-4">Récapitulatif</h2>
 
         {/* AFFICHE LES PROPRIÉTÉS, PAS L'OBJET ENTIER */}
+        <h1>Objet : {JSON.stringify(data)}</h1>
         <p>Email : {data.email}</p>
         <p>Mot de passe : {data.password}</p>
